refactor: extract scope and robots modules from crawl
